enable one vanilla test with --generate-client-interfaces
