+++ vanilla-tests/src/main/java/fixtures/bodystring/Enums.java
@@ -1,373 +1,206 @@
 package fixtures.bodystring;
 
-import com.azure.core.annotation.BodyParam;
-import com.azure.core.annotation.ExpectedResponses;
-import com.azure.core.annotation.Get;
-import com.azure.core.annotation.Host;
-import com.azure.core.annotation.HostParam;
-import com.azure.core.annotation.Put;
 import com.azure.core.annotation.ReturnType;
-import com.azure.core.annotation.ServiceInterface;
 import com.azure.core.annotation.ServiceMethod;
-import com.azure.core.annotation.UnexpectedResponseExceptionType;
 import com.azure.core.http.rest.Response;
-import com.azure.core.http.rest.RestProxy;
-import com.azure.core.util.Context;
-import com.azure.core.util.FluxUtil;
 import fixtures.bodystring.models.Colors;
-import fixtures.bodystring.models.ErrorException;
 import fixtures.bodystring.models.RefColorConstant;
 import reactor.core.publisher.Mono;
 
 /** An instance of this class provides access to all the operations defined in Enums. */
-public final class Enums {
-    /** The proxy service used to perform REST calls. */
-    private final EnumsService service;
-
-    /** The service client containing this operation class. */
-    private final AutoRestSwaggerBATService client;
-
-    /**
-     * Initializes an instance of Enums.
-     *
-     * @param client the instance of the service client containing this operation class.
-     */
-    Enums(AutoRestSwaggerBATService client) {
-        this.service = RestProxy.create(EnumsService.class, client.getHttpPipeline(), client.getSerializerAdapter());
-        this.client = client;
-    }
-
-    /**
-     * The interface defining all the services for AutoRestSwaggerBATServiceEnums to be used by the proxy service to
-     * perform REST calls.
-     */
-    @Host("{$host}")
-    @ServiceInterface(name = "AutoRestSwaggerBATSe")
-    private interface EnumsService {
-        @Get("/string/enum/notExpandable")
-        @ExpectedResponses({200})
-        @UnexpectedResponseExceptionType(ErrorException.class)
-        Mono<Response<Colors>> getNotExpandable(@HostParam("$host") String host, Context context);
-
-        @Put("/string/enum/notExpandable")
-        @ExpectedResponses({200})
-        @UnexpectedResponseExceptionType(ErrorException.class)
-        Mono<Response<Void>> putNotExpandable(
-                @HostParam("$host") String host, @BodyParam("application/json") Colors stringBody, Context context);
-
-        @Get("/string/enum/Referenced")
-        @ExpectedResponses({200})
-        @UnexpectedResponseExceptionType(ErrorException.class)
-        Mono<Response<Colors>> getReferenced(@HostParam("$host") String host, Context context);
-
-        @Put("/string/enum/Referenced")
-        @ExpectedResponses({200})
-        @UnexpectedResponseExceptionType(ErrorException.class)
-        Mono<Response<Void>> putReferenced(
-                @HostParam("$host") String host, @BodyParam("application/json") Colors enumStringBody, Context context);
-
-        @Get("/string/enum/ReferencedConstant")
-        @ExpectedResponses({200})
-        @UnexpectedResponseExceptionType(ErrorException.class)
-        Mono<Response<RefColorConstant>> getReferencedConstant(@HostParam("$host") String host, Context context);
-
-        @Put("/string/enum/ReferencedConstant")
-        @ExpectedResponses({200})
-        @UnexpectedResponseExceptionType(ErrorException.class)
-        Mono<Response<Void>> putReferencedConstant(
-                @HostParam("$host") String host,
-                @BodyParam("application/json") RefColorConstant enumStringBody,
-                Context context);
-    }
-
+public interface Enums {
     /**
      * Get enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<Colors>> getNotExpandableWithResponseAsync() {
-        if (this.client.getHost() == null) {
-            return Mono.error(
-                    new IllegalArgumentException("Parameter this.client.getHost() is required and cannot be null."));
-        }
-        return FluxUtil.withContext(context -> service.getNotExpandable(this.client.getHost(), context));
-    }
+    Mono<Response<Colors>> getNotExpandableWithResponseAsync();
 
     /**
      * Get enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Colors> getNotExpandableAsync() {
-        return getNotExpandableWithResponseAsync()
-                .flatMap(
-                        (Response<Colors> res) -> {
-                            if (res.getValue() != null) {
-                                return Mono.just(res.getValue());
-                            } else {
-                                return Mono.empty();
-                            }
-                        });
-    }
+    Mono<Colors> getNotExpandableAsync();
 
     /**
      * Get enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Colors getNotExpandable() {
-        return getNotExpandableAsync().block();
-    }
+    Colors getNotExpandable();
 
     /**
      * Sends value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
      * @param stringBody string body.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the completion.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<Void>> putNotExpandableWithResponseAsync(Colors stringBody) {
-        if (this.client.getHost() == null) {
-            return Mono.error(
-                    new IllegalArgumentException("Parameter this.client.getHost() is required and cannot be null."));
-        }
-        if (stringBody == null) {
-            return Mono.error(new IllegalArgumentException("Parameter stringBody is required and cannot be null."));
-        }
-        return FluxUtil.withContext(context -> service.putNotExpandable(this.client.getHost(), stringBody, context));
-    }
+    Mono<Response<Void>> putNotExpandableWithResponseAsync(Colors stringBody);
 
     /**
      * Sends value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
      * @param stringBody string body.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the completion.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Void> putNotExpandableAsync(Colors stringBody) {
-        return putNotExpandableWithResponseAsync(stringBody).flatMap((Response<Void> res) -> Mono.empty());
-    }
+    Mono<Void> putNotExpandableAsync(Colors stringBody);
 
     /**
      * Sends value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
      * @param stringBody string body.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public void putNotExpandable(Colors stringBody) {
-        putNotExpandableAsync(stringBody).block();
-    }
+    void putNotExpandable(Colors stringBody);
 
     /**
      * Get enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<Colors>> getReferencedWithResponseAsync() {
-        if (this.client.getHost() == null) {
-            return Mono.error(
-                    new IllegalArgumentException("Parameter this.client.getHost() is required and cannot be null."));
-        }
-        return FluxUtil.withContext(context -> service.getReferenced(this.client.getHost(), context));
-    }
+    Mono<Response<Colors>> getReferencedWithResponseAsync();
 
     /**
      * Get enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Colors> getReferencedAsync() {
-        return getReferencedWithResponseAsync()
-                .flatMap(
-                        (Response<Colors> res) -> {
-                            if (res.getValue() != null) {
-                                return Mono.just(res.getValue());
-                            } else {
-                                return Mono.empty();
-                            }
-                        });
-    }
+    Mono<Colors> getReferencedAsync();
 
     /**
      * Get enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return enum value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Colors getReferenced() {
-        return getReferencedAsync().block();
-    }
+    Colors getReferenced();
 
     /**
      * Sends value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
      * @param enumStringBody enum string body.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the completion.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<Void>> putReferencedWithResponseAsync(Colors enumStringBody) {
-        if (this.client.getHost() == null) {
-            return Mono.error(
-                    new IllegalArgumentException("Parameter this.client.getHost() is required and cannot be null."));
-        }
-        if (enumStringBody == null) {
-            return Mono.error(new IllegalArgumentException("Parameter enumStringBody is required and cannot be null."));
-        }
-        return FluxUtil.withContext(context -> service.putReferenced(this.client.getHost(), enumStringBody, context));
-    }
+    Mono<Response<Void>> putReferencedWithResponseAsync(Colors enumStringBody);
 
     /**
      * Sends value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
      * @param enumStringBody enum string body.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the completion.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Void> putReferencedAsync(Colors enumStringBody) {
-        return putReferencedWithResponseAsync(enumStringBody).flatMap((Response<Void> res) -> Mono.empty());
-    }
+    Mono<Void> putReferencedAsync(Colors enumStringBody);
 
     /**
      * Sends value 'red color' from enumeration of 'red color', 'green-color', 'blue_color'.
      *
      * @param enumStringBody enum string body.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public void putReferenced(Colors enumStringBody) {
-        putReferencedAsync(enumStringBody).block();
-    }
+    void putReferenced(Colors enumStringBody);
 
     /**
      * Get value 'green-color' from the constant.
      *
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return value 'green-color' from the constant.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<RefColorConstant>> getReferencedConstantWithResponseAsync() {
-        if (this.client.getHost() == null) {
-            return Mono.error(
-                    new IllegalArgumentException("Parameter this.client.getHost() is required and cannot be null."));
-        }
-        return FluxUtil.withContext(context -> service.getReferencedConstant(this.client.getHost(), context));
-    }
+    Mono<Response<RefColorConstant>> getReferencedConstantWithResponseAsync();
 
     /**
      * Get value 'green-color' from the constant.
      *
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return value 'green-color' from the constant.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<RefColorConstant> getReferencedConstantAsync() {
-        return getReferencedConstantWithResponseAsync()
-                .flatMap(
-                        (Response<RefColorConstant> res) -> {
-                            if (res.getValue() != null) {
-                                return Mono.just(res.getValue());
-                            } else {
-                                return Mono.empty();
-                            }
-                        });
-    }
+    Mono<RefColorConstant> getReferencedConstantAsync();
 
     /**
      * Get value 'green-color' from the constant.
      *
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return value 'green-color' from the constant.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public RefColorConstant getReferencedConstant() {
-        return getReferencedConstantAsync().block();
-    }
+    RefColorConstant getReferencedConstant();
 
     /**
      * Sends value 'green-color' from a constant.
      *
      * @param enumStringBody enum string body.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the completion.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Response<Void>> putReferencedConstantWithResponseAsync(RefColorConstant enumStringBody) {
-        if (this.client.getHost() == null) {
-            return Mono.error(
-                    new IllegalArgumentException("Parameter this.client.getHost() is required and cannot be null."));
-        }
-        if (enumStringBody == null) {
-            return Mono.error(new IllegalArgumentException("Parameter enumStringBody is required and cannot be null."));
-        } else {
-            enumStringBody.validate();
-        }
-        return FluxUtil.withContext(
-                context -> service.putReferencedConstant(this.client.getHost(), enumStringBody, context));
-    }
+    Mono<Response<Void>> putReferencedConstantWithResponseAsync(RefColorConstant enumStringBody);
 
     /**
      * Sends value 'green-color' from a constant.
      *
      * @param enumStringBody enum string body.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the completion.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public Mono<Void> putReferencedConstantAsync(RefColorConstant enumStringBody) {
-        return putReferencedConstantWithResponseAsync(enumStringBody).flatMap((Response<Void> res) -> Mono.empty());
-    }
+    Mono<Void> putReferencedConstantAsync(RefColorConstant enumStringBody);
 
     /**
      * Sends value 'green-color' from a constant.
      *
      * @param enumStringBody enum string body.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws ErrorException thrown if the request is rejected by server.
+     * @throws fixtures.bodystring.models.ErrorException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    public void putReferencedConstant(RefColorConstant enumStringBody) {
-        putReferencedConstantAsync(enumStringBody).block();
-    }
+    void putReferencedConstant(RefColorConstant enumStringBody);
 }
