+++ vanilla-tests/src/main/java/fixtures/bodystring/implementation/AutoRestSwaggerBATServiceImpl.java
@@ -1,0 +1,114 @@
+package fixtures.bodystring.implementation;
+
+import com.azure.core.http.HttpPipeline;
+import com.azure.core.http.HttpPipelineBuilder;
+import com.azure.core.http.policy.CookiePolicy;
+import com.azure.core.http.policy.RetryPolicy;
+import com.azure.core.http.policy.UserAgentPolicy;
+import com.azure.core.util.serializer.JacksonAdapter;
+import com.azure.core.util.serializer.SerializerAdapter;
+import fixtures.bodystring.AutoRestSwaggerBATService;
+import fixtures.bodystring.Enums;
+import fixtures.bodystring.StringOperations;
+
+/** Initializes a new instance of the AutoRestSwaggerBATService type. */
+public final class AutoRestSwaggerBATServiceImpl implements AutoRestSwaggerBATService {
+    /** server parameter. */
+    private final String host;
+
+    /**
+     * Gets server parameter.
+     *
+     * @return the host value.
+     */
+    public String getHost() {
+        return this.host;
+    }
+
+    /** The HTTP pipeline to send requests through. */
+    private final HttpPipeline httpPipeline;
+
+    /**
+     * Gets The HTTP pipeline to send requests through.
+     *
+     * @return the httpPipeline value.
+     */
+    public HttpPipeline getHttpPipeline() {
+        return this.httpPipeline;
+    }
+
+    /** The serializer to serialize an object into a string. */
+    private final SerializerAdapter serializerAdapter;
+
+    /**
+     * Gets The serializer to serialize an object into a string.
+     *
+     * @return the serializerAdapter value.
+     */
+    public SerializerAdapter getSerializerAdapter() {
+        return this.serializerAdapter;
+    }
+
+    /** The StringOperations object to access its operations. */
+    private final StringOperations stringOperations;
+
+    /**
+     * Gets the StringOperations object to access its operations.
+     *
+     * @return the StringOperations object.
+     */
+    public StringOperations getStringOperations() {
+        return this.stringOperations;
+    }
+
+    /** The Enums object to access its operations. */
+    private final Enums enums;
+
+    /**
+     * Gets the Enums object to access its operations.
+     *
+     * @return the Enums object.
+     */
+    public Enums getEnums() {
+        return this.enums;
+    }
+
+    /**
+     * Initializes an instance of AutoRestSwaggerBATService client.
+     *
+     * @param host server parameter.
+     */
+    AutoRestSwaggerBATServiceImpl(String host) {
+        this(
+                new HttpPipelineBuilder()
+                        .policies(new UserAgentPolicy(), new RetryPolicy(), new CookiePolicy())
+                        .build(),
+                JacksonAdapter.createDefaultSerializerAdapter(),
+                host);
+    }
+
+    /**
+     * Initializes an instance of AutoRestSwaggerBATService client.
+     *
+     * @param httpPipeline The HTTP pipeline to send requests through.
+     * @param host server parameter.
+     */
+    AutoRestSwaggerBATServiceImpl(HttpPipeline httpPipeline, String host) {
+        this(httpPipeline, JacksonAdapter.createDefaultSerializerAdapter(), host);
+    }
+
+    /**
+     * Initializes an instance of AutoRestSwaggerBATService client.
+     *
+     * @param httpPipeline The HTTP pipeline to send requests through.
+     * @param serializerAdapter The serializer to serialize an object into a string.
+     * @param host server parameter.
+     */
+    AutoRestSwaggerBATServiceImpl(HttpPipeline httpPipeline, SerializerAdapter serializerAdapter, String host) {
+        this.httpPipeline = httpPipeline;
+        this.serializerAdapter = serializerAdapter;
+        this.host = host;
+        this.stringOperations = new StringOperationsImpl(this);
+        this.enums = new EnumsImpl(this);
+    }
+}
