+++ vanilla-tests/src/main/java/fixtures/bodystring/implementation/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package containing the implementations for AutoRestSwaggerBATService. Test Infrastructure for AutoRest Swagger BAT.
+ */
+package fixtures.bodystring.implementation;
